+++ src/components/ui/home/PrintPDF/UserList/index.tsx
@@ -1,0 +1,16 @@
+import * as S from './style';
+
+const UserList = () => {
+  return (
+    <S.UserList>
+      <S.UserListItem>
+        <S.UserListText>sj060706</S.UserListText>
+      </S.UserListItem>
+      <S.UserListItem>
+        <S.UserListText>123456789</S.UserListText>
+      </S.UserListItem>
+    </S.UserList>
+  );
+};
+
+export default UserList;
